refactor(analysis): extract derived integrity owner
